+++ sshd-common/src/main/java/org/apache/sshd/common/cipher/BaseRC4Cipher.java
@@ -53,4 +53,9 @@ public class BaseRC4Cipher extends BaseCipher {
 
         return instance;
     }
+
+    @Override
+    public void update(byte[] input, int inputOffset, int inputLen) throws Exception {
+        getCipherInstance().update(input, inputOffset, inputLen, input, inputOffset);
+    }
 }