+++ CHANGES.md
@@ -32,6 +32,8 @@
 
 ## Bug Fixes
 
+* [GH-455](https://github.com/apache/mina-sshd/issues/455) Fix `BaseCipher`: make sure all bytes are processed
+
 ## New Features
 
 ## Behavioral changes and enhancements